@@ -1,53 +1,65 @@
-﻿namespace RefactorMoveFolderValidation {
-    public class MoveFolderValidation {
-        private readonly User currentUser;
-        private readonly ICustomReportFolderService customReportFolderService;
+﻿using System;
 
-        public MoveFolderValidation(
-            User currentUser,
-            ICustomReportFolderService customReportFolderService) {
-            this.currentUser = currentUser;
-            this.customReportFolderService = customReportFolderService;
-        }
+namespace RefactorMoveFolderValidation;
 
-        /// <summary>
-        /// Validates whether the user is permitted to move the item into the target folder.
-        /// </summary>
-        public Result ValidateMoveFolderItem(IFolderItem item, long? targetFolderId) {
-            var result = new Result();
-            var isOwner = item.IsOwner(this.currentUser);
-            var isAdminMovingSharedItem = this.IsAdminMovingSharedItem(item);
+public class MoveFolderValidation
+{
+	private readonly User currentUser;
+	private readonly ICustomReportFolderService customReportFolderService;
 
-            if (!isOwner && !isAdminMovingSharedItem) {
-                result.AddErrorByCode(FolderItemErrorCode.MoveItemOrFolderCreatedByAnotherUser);
-            }
+	public MoveFolderValidation(
+		User currentUser,
+		ICustomReportFolderService customReportFolderService)
+	{
+		this.currentUser = currentUser;
+		this.customReportFolderService = customReportFolderService;
+	}
 
-            bool isAdminMovingToSharedFolder;
-            if (targetFolderId.HasValue) {
-                var targetFolder = this.customReportFolderService.Get(targetFolderId.Value);
+	/// <summary>
+	/// Validates whether the user is permitted to move the item into the target folder.
+	/// </summary>
+	public Result ValidateMoveFolderItem(IFolderItem item, long? targetFolderId)
+	{
+		Result result = new();
+		bool isOwner = item.IsOwner(currentUser);
+		bool isAdminMovingSharedItem = IsAdminMovingSharedItem(item);
 
-                isOwner = targetFolder.IsOwner(this.currentUser);
-                isAdminMovingToSharedFolder = this.IsAdminMovingSharedItem(targetFolder);
+		if (!isOwner && !isAdminMovingSharedItem)
+		{
+			result.AddErrorByCode(FolderItemErrorCode.MoveItemOrFolderCreatedByAnotherUser);
+		}
 
-                if (!isOwner && !isAdminMovingToSharedFolder) {
-                    result.AddErrorByCode(FolderItemErrorCode.MoveItemOrFolderToFolderCreatedByAnotherUser);
-                }
-            } else {
-                // root folder
-                isAdminMovingToSharedFolder = !isOwner && this.currentUser.RoleId == RoleId.Admin;
-            }
+		bool isAdminMovingToSharedFolder;
+		if (targetFolderId.HasValue)
+		{
+			IFolderItem targetFolder = customReportFolderService.Get(targetFolderId.Value);
 
-            if (isAdminMovingSharedItem && !isAdminMovingToSharedFolder) {
-                result.AddErrorByCode(FolderItemErrorCode.SharedItemMayOnlyMoveToSharedFolder);
-            }
+			isOwner = targetFolder.IsOwner(currentUser);
+			isAdminMovingToSharedFolder = IsAdminMovingSharedItem(targetFolder);
 
-            return result;
-        }
+			if (!isOwner && !isAdminMovingToSharedFolder)
+			{
+				result.AddErrorByCode(FolderItemErrorCode.MoveItemOrFolderToFolderCreatedByAnotherUser);
+			}
+		}
+		else
+		{
+			// root folder
+			isAdminMovingToSharedFolder = !isOwner && currentUser.RoleId == RoleId.Admin;
+		}
 
-        private bool IsAdminMovingSharedItem(IFolderItem item) {
-            var isOwner = item.IsOwner(this.currentUser);
-            var isAdminMovingSharedItem = !isOwner && this.currentUser.RoleId == RoleId.Admin && item.IsShared;
-            return isAdminMovingSharedItem;
-        }
-    }
+		if (isAdminMovingSharedItem && !isAdminMovingToSharedFolder)
+		{
+			result.AddErrorByCode(FolderItemErrorCode.SharedItemMayOnlyMoveToSharedFolder);
+		}
+
+		return result;
+	}
+
+	private bool IsAdminMovingSharedItem(IFolderItem item)
+	{
+		bool isOwner = item.IsOwner(currentUser);
+		bool isAdminMovingSharedItem = !isOwner && currentUser.RoleId == RoleId.Admin && item.IsShared;
+		return isAdminMovingSharedItem;
+	}
 }
